requirements.txt
bs4
schedule
Flask
flask
cassandra-driver
bs4
requests
apscheduler
